algunas mejoras

+++ portal-empleado-mysql/controllers/incidences/deleteIncidence.js
@@ -1,27 +1,18 @@
 'use strict';
 
 const { getConnection } = require('../../db');
-const {
-  /*     formatDateToDB,
-    sendEmail,
-    randomString, */
-  generateError
-} = require('../../helpers');
-
-/* const {
-    incidenceSchema,
-    voteSchema,
-    searchSchema,
-    editIncidenceSchema
-} = require('../validations'); */
+const { generateError } = require('../../helpers');
 
 // DELETE - /INCIDENCES/:id
 async function deleteIncidence(req, res, next) {
   let connection;
   try {
     const { id } = req.params;
+    const roleId = req.auth.role;
 
     connection = await getConnection();
+
+    console.log(roleId);
 
     // Delete incidence and vote if exists!
     const [
@@ -31,12 +22,27 @@ async function deleteIncidence(req, res, next) {
     if (!current.length) {
       throw generateError(`No hay ninguna incidencia con el número ${id}`, 400);
     }
+    const [data] = current;
+    if (roleId === 'admin') {
+      await connection.query(
+        'DELETE FROM valoraciones WHERE incidencias_id=?',
+        [id]
+      );
+      await connection.query('DELETE FROM incidencias WHERE id=?', [id]);
+    } else {
+      if (data.activo === 0) {
+        throw generateError(
+          `la incidencia número ${id} no se puede borrar, ya ha sido gestionada`,
+          409
+        );
+      }
 
-    await connection.query('DELETE FROM valoraciones WHERE incidencias_id=?', [
-      id
-    ]);
-    await connection.query('DELETE FROM incidencias WHERE id=?', [id]);
-
+      await connection.query(
+        'DELETE FROM valoraciones WHERE incidencias_id=?',
+        [id]
+      );
+      await connection.query('DELETE FROM incidencias WHERE id=?', [id]);
+    }
     res.send({
       status: 'ok',
       message: `La incidencia con el número ${id} ha sido borrada`
